enqueue visible cells/headers/footers in reverse order
so they will be dequeued in their original order.  Helpful in situations
where reloadData() is called, but many/all cells stay the same.  Otherwise
we see some flickering as items are placed in the wrong location and then
moved quickly.  NOTE: it's unclear why the flickering occurs, but it's
likely somethig in UIDisplayServer/UIHTMLDisplayServer that still needs to be fixed.
Until then, this new queue ordering helps minimize the visual flickering because
unchanged items never move