redfine JSData to be an alias for Uint8Array
Previously, JSData wrapped a Uint8Array.

I cannot remember the original reason for this wrapping, and can find no
issue with going ahead with the redifinition.

The wrapping caused a lot of unnecessary code to either wrap or unwrap
a Uint8Array, making larger blocks of code hard to read.

Uint8Array is such a silly name, though, that I'm keeping JSData as an
alias.  All of the JSData extension methods are still there, but now are
on any Uint8Array.  There are also JSData init methods that I find more
clear and easier to use than the Uint8Array constructors.

Note for native node code uses Buffer, which is like, but not identical to
Uint8Array.  When receiving a Buffer from node calls, it should be remade
into a JSData/Uint8Array via JSData.initWithNodeBuffer(buffer)